Adapt File-structure
