+++ learners/__init__.py
@@ -31,9 +31,21 @@ def main():
 
         init_mail(app)
 
-    from learners import views
+    from learners.routes.home import home_api
+    from learners.routes.authentication import authentication_api
+    from learners.routes.interface import interface_api
+    from learners.routes.execution import execution_api
+    from learners.routes.callback import callback_api
+    from learners.routes.statics import statics_api
+    from learners.routes.admin import admin_api
 
-    app.register_blueprint(views.bp)
+    app.register_blueprint(home_api)
+    app.register_blueprint(authentication_api)
+    app.register_blueprint(interface_api)
+    app.register_blueprint(execution_api)
+    app.register_blueprint(callback_api)
+    app.register_blueprint(statics_api)
+    app.register_blueprint(admin_api)
 
     return app
