+++ learners/database.py
@@ -1,15 +1,9 @@
-import json
-import os
-import pathlib
+# sourcery skip: avoid-builtin-shadow
 from flask_sqlalchemy import SQLAlchemy
-from datetime import datetime
+from sqlalchemy import event, func
+
 from learners.conf.config import cfg
-
-from bs4 import BeautifulSoup
-
-from sqlalchemy import event
-from sqlalchemy import func
-
+from learners.functions.helpers import get_exercises
 
 """
 Set up the database
@@ -24,8 +18,6 @@ db = SQLAlchemy()
 class User(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     username = db.Column(db.String(20), unique=True, nullable=False)
-    scriptExercises = db.relationship("ScriptExercise", backref="user", lazy=True)
-    formExercises = db.relationship("FormExercise", backref="user", lazy=True)
     executions = db.relationship("Execution", backref="user", lazy=True)
 
 
@@ -60,29 +52,6 @@ class TokenBlocklist(db.Model):
     created_at = db.Column(db.DateTime, nullable=False)
 
 
-# History of sent POSTs
-class ScriptExercise(db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    script_name = db.Column(db.String(120), nullable=False)
-    call_uuid = db.Column(db.String(120), unique=True, nullable=False)
-    start_time = db.Column(db.DateTime, nullable=False, default=datetime.utcnow)
-    response_time = db.Column(db.DateTime, nullable=True)
-    response_content = db.Column(db.Text, nullable=True)
-    completed = db.Column(db.Integer, nullable=False, default=0)
-    connection_failed = db.Column(db.Integer, nullable=False, default=0)
-    msg = db.Column(db.String(240), nullable=True)
-    user_id = db.Column(db.Integer, db.ForeignKey("user.id"), nullable=False)
-
-
-# Formdata
-class FormExercise(db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    name = db.Column(db.String(120), nullable=False)
-    data = db.Column(db.String(), nullable=False)
-    timestamp = db.Column(db.DateTime, nullable=False, default=datetime.utcnow)
-    user_id = db.Column(db.Integer, db.ForeignKey("user.id"), nullable=False)
-
-
 @event.listens_for(User.__table__, "after_create")
 def insert_initial_users(*args, **kwargs):
     for user, _ in cfg.users.items():
@@ -108,39 +77,7 @@ def insert_exercises(*args, **kwargs):
     db.session.commit()
 
 
-def get_exercises():
-    exercises = [{"id": "all", "type": "all", "exerciseWeight": 0, "parentWeight": "0", "name": "all"}]
-
-    if cfg.exercises.get("directory").startswith("/"):
-        root_directory = f"{cfg.exercises.get('directory')}/{list(cfg.users.keys())[0]}/en/"
-    else:
-        root_directory = f"./learners/{cfg.exercises.get('directory')}/{list(cfg.users.keys())[0]}/en/"
-
-    for path, subdirs, files in os.walk(root_directory):
-        for file in files:
-            if file.endswith(".html"):
-                f = open(pathlib.PurePath(path, file), "r")
-                parsed_html = BeautifulSoup(f.read(), features="html.parser")
-                exerciseInfos = parsed_html.body.find_all("input", attrs={"class": "exercise-info"})
-                exercises.extend(json.loads(exerciseInfo.get("value")) for exerciseInfo in exerciseInfos)
-
-    for exercise in exercises:
-        exerciseWeight = int(exercise["exerciseWeight"])
-        parentWeight = int(exercise["parentWeight"])
-
-        if parentWeight == 0:
-            exercise["exerciseWeight"] = exerciseWeight * 10
-        else:
-            exercise["exerciseWeight"] = parentWeight * 10 + exerciseWeight
-
-        exercise["name"] = exercise["id"].replace("_", " ")
-
-    exercises = sorted(exercises, key=lambda d: d["exerciseWeight"])
-    return exercises
-
-
 def build_db(app):
     global db
     db.init_app(app)
-
     db.create_all()
