+++ learners/functions/authentication.py
@@ -1,0 +1,2 @@
+def check_password(usermap: dict, user: str, password: str) -> bool:
+    return user in usermap and usermap.get(user).get("password") == password
